+++ app/schemas/__init__.py
@@ -1,4 +1,18 @@
 from .users import User, Users, UserType, UserSearch
 from .point import PointHistory, PointHistoryType
+from .quest import Quest, QuestCompletion
+from .post import Post, Posts, PostContent
 
-__all__ = ["User", "Users", "UserType", "UserSearch", "PointHistory", "PointHistoryType"]
+__all__ = [
+    "User",
+    "Users",
+    "UserType",
+    "UserSearch",
+    "PointHistory",
+    "PointHistoryType",
+    "Quest",
+    "QuestCompletion",
+    "Post",
+    "Posts",
+    "PostContent"
+]